feat: Configure API URL during Docker build
This commit configures the API URL during the Docker build process by passing it as a build argument.

The `VITE_API_URL` build argument is now set to `http://flamware.work/api` during the `docker build` command. This allows the application to correctly connect to the API server when running in a Docker container.